+++ Assets/FlashTools/Docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ###### Version 1.3.15
 * Add bitmap trimming
 * Fix preview leaks in the Editor mode
+* Add warning notes about outdated assets
 
 ###### Version 1.3.14
 * Fix 2018.3.2f1 compilation